+++ src/main/kotlin/icu/windea/pls/images/dds/DdsFileType.kt
@@ -9,9 +9,13 @@ import icu.windea.pls.PlsBundle
 object DdsFileType : UserBinaryFileType() {
     override fun getName() = "DDS"
 
+    override fun getDefaultExtension(): String = "dds"
+
     override fun getDisplayName() = PlsBundle.message("filetype.dds.display.name")
 
     override fun getDescription() = PlsBundle.message("filetype.dds.description")
 
     override fun getIcon() = AllIcons.FileTypes.Image
+
+    val contentTypeString get() = "description=${description};file_extensions=${defaultExtension}"
 }