+++ src/main/kotlin/icu/windea/pls/images/tga/TgaFileType.kt
@@ -9,9 +9,13 @@ import icu.windea.pls.PlsBundle
 object TgaFileType : UserBinaryFileType() {
     override fun getName() = "TGA"
 
+    override fun getDefaultExtension(): String = "tga"
+
     override fun getDisplayName() = PlsBundle.message("filetype.tga.display.name")
 
     override fun getDescription() = PlsBundle.message("filetype.tga.description")
 
     override fun getIcon() = AllIcons.FileTypes.Image
+
+    val contentTypeString get() = "description=${description};file_extensions=${defaultExtension}"
 }